changed: wavelength is now a double

+++ gui/default_parameters_widget.cpp
@@ -365,7 +365,7 @@ bool DefaultParametersWidget::getWavelengthPattern() const
 }
     
 
-int DefaultParametersWidget::getWavelength() const
+double DefaultParametersWidget::getWavelength() const
 {
     return 0;
 }
